Change to timestamp type to uint64

+++ firmware/dwm1001_dev_source_code/development/dedicated_time_protocol/main.h
@@ -4,6 +4,6 @@
 void syncCycle(void);
 void rxHandler(msg_template *msg);
 void configTx2(void);
-void updateTx1Ts(uint32 ts);
+void updateTx1Ts(uint64 ts);
 void setRxTimeout2(void);
 void resetRxTimeout(uint8 id);
